Add feature to allow player to choose color
Because the final goal is to create a one-player chess game, I added a feature to allow the player to choose whether they wanted to be in control over the white or black pieces. The player's pieces will always be positioned at the lower section of the board. In order to add this feature, I had to modify the code in ChessPieces.js that creates the players, code in possibleMoves() function of ChessPieceLogic.js, and also add another component (GameButtons) that has buttons to allow the player to choose which color piece they want to be in control of.